+++ plugin/src/main/kotlin/tech/skot/Versions.kt
@@ -7,21 +7,21 @@ import kotlin.String
  * This code is generated from build.gradle.kts
  */
 public object Versions {
-  public const val skot: String = "1.3.3-ua"
+  public const val skot: String = "1.3.4-ua"
 
   public const val group: String = "com.github.useradgents.skot"
 
-  public const val serialization: String = "1.7.1"
+  public const val serialization: String = "1.7.3"
 
-  public const val kotlinxDateTime: String = "0.6.0"
+  public const val kotlinxDateTime: String = "0.6.1"
 
   public const val ktor: String = "2.3.12"
 
-  public const val kotlin: String = "2.0.0"
+  public const val kotlin: String = "2.0.20"
 
-  public const val kotlinCoroutines: String = "1.8.1"
+  public const val kotlinCoroutines: String = "1.9.0"
 
-  public const val kotlinpoet: String = "1.17.0"
+  public const val kotlinpoet: String = "1.18.1"
 
   public const val android_minSdk: Int = 23
 
@@ -32,4 +32,10 @@ public object Versions {
   public const val android_app_compat: String = "1.7.0"
 
   public const val sqldelight: String = "1.5.5"
+
+  public const val desugaring: String = "2.1.2"
+
+  public const val leakcanary: String = "2.14"
+
+  public const val ktlint: String = "1.3.1"
 }