+++ plugin/src/main/kotlin/tech/skot/tools/gradle/PluginTools.kt
@@ -125,6 +125,6 @@ class PluginTools : Plugin<Project> {
         this.add("implementation", project("$parentProjectPath:viewcontract"))
         this.add("implementation", project("$parentProjectPath:modelcontract"))
         this.add("api", "${Versions.group}:generator:${Versions.skot}")
-        this.add("implementation", "com.pinterest.ktlint:ktlint-cli:1.1.0")
+        this.add("implementation", "com.pinterest.ktlint:ktlint-cli:${Versions.ktlint}")
     }
 }